streamline draw_multilayer function
